requirements.txt
google-cloud-bigquery
google-auth
google-api-python-client
langchain
langchain-google-genai
